Upgrade Cucumber dependency
There was a warning that the notation '~@tag' is deprecated and
should be replaced with 'not @tag'. The code was adapted accordingly.

This version improves HTML report file generation. There are no more
report files per scenerio, only per package and test class. This fixes
the problem that report generation failed under Windows. This was because
scenarios with their parameters contain e.g. pipes and these are no valid
characters for filenames under Windows.