Upgrade parent pom & Gradle
The newer parent pom requires are newer Gradle. Therefore both
of them have to be upgraded at the same time.

Differences due to the new Gradle version:

1) The syntax to configure plugins.

2) The syntax to configure the Gradle wrapper has changed.

3) The syntax to add new tasks has changed.

4) No need to optimize memory via `DEFAULT_JVM_OPTS` anymore.
The wrapper is now already configured to only use 64 MB.

Differences due to the new jpi plugin version:

1) Setting `systemProp.jenkins.install.runSetupWizard` in gradle.properties
is not required anymore. The jpi plugin now takes of disabling the
setup wizard itself when `gradle server` is run.

2) The manifest attribute `PluginFirstClassLoader` can now more easily be
configured via the `jenkinsPlugin` block.

3) The `jpi` task no longer extends the `war` task, now it just depends on it.
Therefore extending the classpath to add the Groovy JAR to the HPI file
has to be configured on the `war` task now.

4) The `junit-dep` dependency is no longer added to the `jenkinsTest`
configuration for Jenkins >= 1.505. Therefore it is not required anymore
to exclude it.

5) Setting `jenkins.httpPort` is no longer required as 8080 is the default.

6) Setting repositories is no longer required. They are already set by
the jpi plugin.